requirements.txt
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
PyPDF2
streamlit
